Create requirements.txt
Added a basic requirements file. Its somewhat difficult now to get this project running as installing the latest dependencies leads to package conflicts when pulling the latest version from pypi. I propose adding this as a minimal set of versioned dependencies to make it simpler for those looking to fine-tune CodeT5+ following the guide.